Robustness fixes. Implemented --maxtries and --wait options.

+++ tools/install-dispatcher.sh
@@ -5,7 +5,11 @@ BINDIR=/usr/local/bin
 
 mkdir -p ${DESTDIR}
 cp target/Dispatcher-*-jar-with-dependencies.jar ${DESTDIR}
-cp src/main/config/dispatcher.sh ${DESTDIR}
+if [ ! -e ${DESTDIR}/dispatcher.sh ] ; then
+    cp src/main/config/dispatcher.sh ${DESTDIR}
+else
+    echo "NOT overwriting existing ${DESTDIR}/dispatcher.sh"
+fi
 ln -sf ${DESTDIR}/dispatcher.sh ${BINDIR}/dispatcher
 
 if [ ! -e ${DESTDIR}/listener.xml ] ; then
